Updated indexing for tanh lookup table version and updated tests for the example

+++ hls4ml/templates/catapult/nnet_utils/nnet_activation_stream.h
@@ -473,7 +473,7 @@ void tanh(ac_channel<data_T> &data, ac_channel<res_T> &res) {
         #pragma hls_unroll
         TanHPackLoop: for (int j = 0; j < res_T::size; j++) {
             //#pragma HLS UNROLL
-            int data_round = (int)in_data[j].to_double()*(int)CONFIG_T::table_size/8;
+            int data_round = in_data[j].to_double()*(int)CONFIG_T::table_size/8;
             int index = data_round + 4*(int)CONFIG_T::table_size/8;
             if (index < 0)   index = 0;
             else if (index > CONFIG_T::table_size-1) index = (int)CONFIG_T::table_size-1;
